Resolving outstanding issues, cleaning up final discrepencies
