+++ module/compiler/src/transformer/manager.ts
@@ -15,11 +15,20 @@ export class TransformerManager {
 
     for (const name of ScanApp.findFiles('.ts', x => /support\/transformer[.].*?[.]ts$/.test(x))) {
       const { transformers } = require(name.file);
-      allTransformers.push(...(transformers || []));
-      for (const transformer of (transformers || [])) {
-        console.debug('Configured Transformers', name.module, transformer);
-      }
+      allTransformers.push(...(transformers || []).map((x: any) => { x.file = name.file; return x; }));
     }
+
+    console.debug('Transformers',
+      ...allTransformers.map(x => {
+        const name = (x as any).file.split('node_modules/').pop()!.replace('/support/transformer.', ':').replace('.ts', '');
+        const flags = [
+          ...(x.all ? ['all'] : []),
+          ...(x.before ? ['before'] : []),
+          ...(x.after ? ['after'] : [])
+        ];
+        return `\n\t[${x.type}] ${name} - ${flags.join(' ')}`;
+      })
+    );
 
     this.visitor = new VisitorFactory(allTransformers);
     this.transformers = {
