+++ module/di/bin/lib/index.ts
@@ -1,0 +1,8 @@
+import { AppListUtil } from './app-list';
+
+export * from './app-list';
+export * from './run';
+export * from './util';
+
+// For plugin
+export const getAppList = AppListUtil.getList.bind(AppListUtil);
