+++ module/di/bin/travetto-plugin-run.js
@@ -1,3 +1,3 @@
 require('@travetto/boot/bin/init')
-  .libRequire('@travetto/di/bin/lib')
+  .libRequire('@travetto/di/bin/lib/run')
   .RunUtil.runDirect();
